+++ examples/cloudflare-todomvc/src/live-store-client-do.ts
@@ -1,6 +1,7 @@
 import { DurableObject } from 'cloudflare:workers'
 import type { AlarmInvocationInfo } from '@cloudflare/workers-types'
 import { type ClientDoWithRpcCallback, createStoreDoPromise } from '@livestore/adapter-cloudflare'
+import type { CfTypes } from '@livestore/sync-cf/cf-worker'
 import { nanoid, type Store, type Unsubscribe } from '@livestore/livestore'
 import { handleSyncUpdateRpc } from '@livestore/sync-cf/client'
 import { schema, tables } from './livestore/schema.ts'
@@ -38,10 +39,12 @@ export class LiveStoreClientDO extends DurableObject<Env> implements ClientDoWit
       storeId,
       clientId: 'client-do',
       sessionId: nanoid(),
-      durableObjectId: this.ctx.id.toString(),
-      bindingName: 'CLIENT_DO',
-      storage: this.ctx.storage,
-      syncBackendDurableObject: this.env.SYNC_BACKEND_DO.get(this.env.SYNC_BACKEND_DO.idFromName(storeId)),
+      durableObject: {
+        state: this.ctx as CfTypes.DurableObjectState,
+        env: this.env,
+        bindingName: 'CLIENT_DO',
+      },
+      syncBackendStub: this.env.SYNC_BACKEND_DO.get(this.env.SYNC_BACKEND_DO.idFromName(storeId)),
       livePull: true,
     })